+++ src/topology.rs
@@ -26,11 +26,11 @@ pub async fn connect_nodes<T: Pea2Pea>(nodes: &[T], topology: Topology) -> io::R
     match topology {
         Topology::Line | Topology::Ring => {
             for i in 0..(count - 1) {
-                let addr = nodes[i + 1].node().listening_addr()?;
+                let addr = nodes[i + 1].node().listening_addr().await?;
                 nodes[i].node().connect(addr).await?;
             }
             if topology == Topology::Ring {
-                let addr = nodes[0].node().listening_addr()?;
+                let addr = nodes[0].node().listening_addr().await?;
                 nodes[count - 1].node().connect(addr).await?;
             }
         }
@@ -39,14 +39,14 @@ pub async fn connect_nodes<T: Pea2Pea>(nodes: &[T], topology: Topology) -> io::R
             for i in 0..count {
                 for (j, peer) in nodes.iter().enumerate() {
                     if i != j && connected_pairs.insert((i, j)) && connected_pairs.insert((j, i)) {
-                        let addr = peer.node().listening_addr()?;
+                        let addr = peer.node().listening_addr().await?;
                         nodes[i].node().connect(addr).await?;
                     }
                 }
             }
         }
         Topology::Star => {
-            let hub_addr = nodes[0].node().listening_addr()?;
+            let hub_addr = nodes[0].node().listening_addr().await?;
             for node in nodes.iter().skip(1) {
                 node.node().connect(hub_addr).await?;
             }