+++ tests/messaging.rs
@@ -91,7 +91,7 @@ async fn messaging_example() {
     let shouter = crate::test_node!("shout");
     shouter.enable_reading().await;
     shouter.enable_writing().await;
-    shouter.node().start_listening().await.unwrap();
+    shouter.node().toggle_listener().await.unwrap().unwrap();
 
     let picky_echo_config = Config {
         name: Some("picky_echo".into()),
@@ -104,7 +104,7 @@ async fn messaging_example() {
     picky_echo.enable_reading().await;
     picky_echo.enable_writing().await;
 
-    let picky_echo_addr = picky_echo.node().start_listening().await.unwrap();
+    let picky_echo_addr = picky_echo.node().toggle_listener().await.unwrap().unwrap();
 
     shouter.node().connect(picky_echo_addr).await.unwrap();
 
@@ -140,7 +140,7 @@ async fn messaging_example() {
 async fn drop_connection_on_invalid_message() {
     let reader = crate::test_node!("reader");
     reader.enable_reading().await;
-    let reader_addr = reader.node().start_listening().await.unwrap();
+    let reader_addr = reader.node().toggle_listener().await.unwrap().unwrap();
 
     let writer = crate::test_node!("writer");
     writer.enable_writing().await;
@@ -168,7 +168,7 @@ async fn drop_connection_on_invalid_message() {
 async fn drop_connection_on_zero_read() {
     let reader = crate::test_node!("reader");
     reader.enable_reading().await;
-    let reader_addr = reader.node().start_listening().await.unwrap();
+    let reader_addr = reader.node().toggle_listener().await.unwrap().unwrap();
 
     let peer = crate::test_node!("peer");
 
@@ -193,7 +193,7 @@ async fn drop_connection_on_zero_read() {
 #[tokio::test]
 async fn no_reading_no_delivery() {
     let reader = crate::test_node!("defunct reader");
-    let reader_addr = reader.node().start_listening().await.unwrap();
+    let reader_addr = reader.node().toggle_listener().await.unwrap().unwrap();
 
     let writer = crate::test_node!("writer");
     writer.enable_writing().await;
@@ -224,7 +224,7 @@ async fn no_reading_no_delivery() {
 async fn no_writing_no_delivery() {
     let reader = crate::test_node!("reader");
     reader.enable_reading().await;
-    let reader_addr = reader.node().start_listening().await.unwrap();
+    let reader_addr = reader.node().toggle_listener().await.unwrap().unwrap();
 
     let writer = crate::test_node!("defunct writer");